requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.2
idna==3.2
pushbullet.py==0.12.0
python-magic==0.4.24
requests==2.26.0
urllib3==1.26.6
websocket-client==1.1.0
